Refactored character position calculations and text processing in data interface
Introduced an iterator pattern that encapsulates the "skip \r\n" logic and used that throughout the class for consistency. Reduced position calculations from 5 passes with 2d arrays involved, to just 2 with flat arrays. Also added support for kerning and vertical offset. Added more features and the line count is lower so must be a good change.