Reverted plugin back to 5.3
Didn't realize that assets aren't backwards compatible, but are forwards compatible. Changing version to 5.3 makes it compatible with all engine versions after 5.3.

Remade all the assets by hand in 5.3

Had to get rid of an optimization in the Data Interface since GetSimTarget doesn't exist in 5.3. It isn't really that big of a deal because it only effects GPU sims which are already slow anyways.